feat: Phase 2 — driver protocols, sequence runner, factor calculation
   - IValveDriver / ISensorSource protocols + HAEntityDriver / HAEntitySensorSource
   - SequenceRunner with asyncio mutex, pause/resume snapshot, watchdog
   - compute_factors: temperature linear factor, rain linear interpolation
   - LiterTracker: records external valve activity as RunHistory
   - 26 passing tests (factors, resume, sequences)
   - config: basis_min_per_zone and range changed to float

+++ src/backend/naiad/config.py
@@ -63,8 +63,8 @@ class ScheduleConfig(BaseModel):
 class SequenceConfig(BaseModel):
     label: str
     zones: list[str]
-    basis_min_per_zone: int
-    range: tuple[int, int] = (5, 240)
+    basis_min_per_zone: float
+    range: tuple[float, float] = (5.0, 240.0)
     watchdog_min: int
     schedule: ScheduleConfig
     enabled: bool = True
